Updated create meetup record api route
This commit modifies project folder and also update package.json
file with npm run eslint,added new coverals.yml,added travis.yml file
,re-factor id generator function to a new file generateID.js and update
gitignore file.

	List of updated file
		* package.json file
		* meeup controller file

	List of new added file
		* travis.yml
		* coveralls.yml
