+++ .coveralls.yml
@@ -1,0 +1,1 @@
+repo_token: h90VNa1p8lUKCNx2WxlPCDwxpIfQzIonr
