+++ .travis.yml
@@ -1,0 +1,11 @@
+language: node_js
+node_js: 
+  - "10.14.2"
+
+
+  cache:
+    directories:
+    - "node_modules"
+after_success:
+    - npm run coveralls
+    - npm run coverage
