+++ app/app.js
@@ -1,15 +1,8 @@
-/**
- *  Import external modules
- */
 import express from 'express';
 import bodyParser from 'body-parser';
 import cors from 'cors';
 import validate from './validations/validate';
 import schema from './validationSchemas/validationSchema';
-
-/**
- *  Import custom routes
- */
 import meetup from './routes/api/meetups/meetup';
 
 // Initialize express app
@@ -25,9 +18,7 @@ app.use(cors());
 // Meet up  middleware
 app.use('/api/v1/', validate.validateBody(schema.createMeetup), meetup);
 
-/**
- * Home page route
- */
+// Home page route
 app.get('/', (req, res) => {
   res.json(
     {
@@ -39,9 +30,7 @@ app.get('/', (req, res) => {
   );
 });
 
-/**
- * Handle non exist route with appropiate message
- */
+// Handle non exist route with with proper message
 app.all('*', (req, res) => {
   res.status(400).json({
     status: 404,
