+++ app/controllers/MeetupController.js
@@ -1,11 +1,8 @@
 // import external modules
 import fs from 'fs';
 import MeetupRecord from '../data/meetuprecord';
+import idGenerator from '../helper/generateID';
 
-/**
- * MeetupController object export
- *
- */
 const MeetupController = {
   /**
    *
@@ -17,27 +14,11 @@ const MeetupController = {
    * @returns object meetup object
    */
   async create(req, res) {
-    // generate unique id
-    const generateUUID = () => {
-      let d = new Date().getTime();
-      if (typeof performance !== 'undefined' && typeof performance.now === 'function') {
-        d += performance.now(); // use high-precision timer if available
-      }
-      return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, (c) => {
-        const r = (d + Math.random() * 16) % 16 | 0;
-        // eslint-disabled "no-bitwise", "no-mixed-operators"
-        d = Math.floor(d / 16);
-        // eslint-disabled "no-bitwise", "no-mixed-operators"
-        return (c === 'x' ? r : (r & 0x3 | 0x8)).toString(16);
-      });
-    };
-
     // Set default value for optional fields if not set
     const images = req.value.body.images || '';
     const tag = req.value.body.tags || '';
     const dateCreated = new Date().toUTCString();
-    const uniqueID = generateUUID();
-
+    const uniqueID = idGenerator();
 
     // get all post request body data
     const values = {
